log file system upgrade

+++ extract_compressed.py
@@ -15,14 +15,18 @@ def main():
     print("-"*24)
     print("\n>>>>>>압축파일 해제<<<<<<\n")
     print("-"*24)
-    folder_path = input("폴더 경로를 입력하세요 (종료는 0을 입력, log파일 삭제는 9를 입력) : ")
+    folder_path = input(
+        "폴더 경로를 입력하세요 (종료는 0을 입력, log파일 삭제는 9를 입력)\n(위원회별로 log를 삭제하는 것을 권장합니다) : ")
 
     if folder_path == '0':
         return 0
 
     if folder_path == '9':
-        os.remove('./log/zip_file.txt')
-        print("로그 파일이 정상적으로 삭제되었습니다")
+        if os.path.exists('./log/zip_file.txt'):
+            os.remove('./log/zip_file.txt')
+            print("로그 파일이 정상적으로 삭제되었습니다")
+        else:
+            print("로그 파일이 존재하지 않습니다.")
         return main()
 
     if not os.path.isdir(folder_path):
